Move CronEvents to Bolt\Events namespace
